+++ changelog.md
@@ -34,6 +34,7 @@
 * LLT-5038: Add doc rs as in-repo documentation system skeleton
 * LLT-5034: Upgrade rust toolchain to v1.77.2
 * LLT-4924: Enable size reduction compiler flags
+* LLT-5078: Fix telio-proxy panic
 
 <br>